First run of changes
HEADER

I changed navigation following the suggestions I made in my previous
report: About/Search/Browse/Contact. I moved all the About content into
a single about page and deleted the other .php files.

I didn't work a lot on design, but I did choose a custom font for the
title and styled a few things.

HOME

I moved a few things around on the home page and added some text to
call attention to the browse feature. I left partners on the home page,
but they are also listed on the about page. I moved the search
suggestions to browse.

ABOUT

All the about info can go here. If the page gets long, we can add some
anchor links at the top to the different sections, but I don't think it
will be necessary.

SEARCH

I added a search.php page and added fake HTML of the advanced search
form. search.php is a duplicate of advanced.php

BROWSE

I remember that browse was dependent on time and what the database
allows - to start, just the list of featured searches and a "Browse all
Documents" would go a long way to adding a bit of usability to the
site. A couple of different sorts would also be nice.

OTHER CHANGES

I moved all the CSS into one file and removed duplicate style
rules-this should make styles easier to change and maintain. On my
site, There is a "flashing" of the website when you click over to a new
page (it shows the page without CSS, and then with). I'm not sure if
this is because my server is slow or what. Let me know if I need to
look into this further.